+++ extensions/healthie/actions/createMetricEntry/index.ts
@@ -1,0 +1,1 @@
+export { createMetricEntry } from './createMetricEntry'