+++ extensions/healthie/validation/createMetricEntry.zod.ts
@@ -1,0 +1,7 @@
+import { z } from 'zod'
+
+export const fieldsValidationSchema = z.object({
+  userId: z.string().nonempty(),
+  category: z.string().nonempty(),
+  metricStat: z.string().nonempty(),
+})